requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
ale-py==0.8.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Box2D==2.3.2
box2d-py==2.3.5
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.6.0
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
datasets==2.20.0
decorator==4.4.2
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
email_validator==2.2.0
etils==1.5.2
Farama-Notifications==0.0.4
fastapi==0.111.1
fastapi-cli==0.0.4
fasteners==0.19
ffmpy==0.3.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.32.0
google-auth-oauthlib==0.4.6
gradio==4.38.1
gradio_client==1.1.0
grpcio==1.37.0
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.1
mujoco==3.1.6
mujoco-py==2.1.2.14
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.6
pandas==2.2.2
parl==2.2.1
pillow==10.4.0
pipenv==2024.0.1
proglog==0.1.10
protobuf==3.20.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.6.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynvml==11.5.3
PyOpenGL==3.1.7
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==22.3.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.4
safetensors==0.4.3
scipy==1.13.1
semantic-version==2.10.0
sentry-sdk==2.10.0
setproctitle==1.3.3
shellingham==1.5.4
Shimmy==0.2.1
swig==4.2.1
sympy==1.13.0
tbb==2021.13.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==2.4.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
virtualenv==20.26.3
wandb==0.17.4
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4

